refactor: move non-reused components to views
